requirements.txt
Flask==3.1.2
requests==2.32.5
beautifulsoup4==4.14.2
openai==2.1.0
python-dotenv==1.1.1
